remove dupe file
